refactor: 폴더정리 및 DiaryDetailView 레이아웃 수정
- MVC로 폴더정리
- DiaryDetailView 레이아웃을 readableContentGuide에 적용
